pyproject.toml
[project]
name = "skinpy"
version = "0.0.1"
description = ""
authors = [
    {name = "Tim Martin", email = "tim@timmart.in"},
]
dependencies = [
    "attrs>=23.1.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "click>=8.1.7",
]
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.292",
    "black>=23.9.1",
    "pytest>=7.4.0",
    "colorspacious>=1.1.2",
    "matplotlib>=3.8.0",
]

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"

[tool.pyright]
plugins = "numpy.typing.mypy_plugin"
